feat(登录): 支持邮箱登录

+++ src/main/java/com/constantineqaq/gateway/utils/UserJwtUtil.java
@@ -31,7 +31,7 @@ public class UserJwtUtil {
      * @param user 用户信息
      * @return 令牌
      */
-    public String createJwt(UserDetails user, String username, Integer userId) {
+    public String createJwt(UserDetails user, String username, Long userId) {
         if(jwtUtil.frequencyCheck(userId)) {
             Algorithm algorithm = Algorithm.HMAC256(jwtUtil.getKey());
             Date expire = jwtUtil.expireTime();
